Adjust `.article-grid` spacing and margins
Reduced the `gap` property in the `.article-grid` class from `20rem` to `2rem` to decrease grid item spacing. Changed `margin-left` from `30px` to `2rem` for consistency with relative units and reduced size. Added a `margin-right` property with a value of `260px` to introduce a right margin, refining the overall layout and alignment.